Don't update text input elements when any attribute is changed
This change removes an unnecessary update to children of
<input type="text"> elements when any attribute is modified on the
element.

A similar patch was made in WebKit here:
https://bugs.webkit.org/show_bug.cgi?id=111990

In the WebKit patch, the update still occurs if the attribute is "value"
or "placeholder." We already handle "value" attribute changes in a case
in HTMLInputElement::ParseAttribute, and we already handle "placeholder"
attribute changes in TextControlElement::ParseAttribute. Since we handle
those cases explicitly, we can remove the entire AttributeChanged
function.

Bug: 1012774
Change-Id: Ib75fde9c5b3e1f9eea4232bf5bfa915d573b9c7e